requirements.txt
matplotlib
trimesh
torch
tqdm
pandas
libigl
opencv-python
open3d
usd-core
transformers==4.55.0
accelerate
qwen-vl-utils
accelerate
vllm
pyglet<2
PySide2
simple_3dviz==0.4.0
seaborn
datasets==3.0
filelock
timm
bitsandbytes
qdrant-client
fastembed-gpu
kaleido==0.2.1
plotly[kaleido]
rtree
openai==0.28